// src/app/model/trpayment.ts
export class trpayment {
    paymentId: number;
    financeNumber: string;
    debtCollectionNumber: string;
    idcard: string;
    deadlineDate: string;
    paymentMethod: string;
    tatalAmountDetail: string;
    totalAmount: number;
    paymentAmount: number;
    balance: number;
    attachmentFile: string;
    status: string;
    isCloseInvoice: string;
    active: string;
    remark: string;
    createdBy: string;
    createDate: Date;
    updatedBy: string;
    updateDate: Date;
    sourceId: string;
    source: string;
};